requirements.txt
# Core LangChain and LangGraph dependencies
langchain>=0.1.0
langgraph>=0.0.32
langchain-google-genai>=0.0.8

# Google AI dependencies
google-generativeai>=0.3.2
google-cloud-aiplatform>=1.38.1

# GUI dependencies
customtkinter>=5.2.0

# Utility dependencies
python-dotenv>=1.0.0
pydantic>=2.5.2
pydantic-settings>=2.0.0
typing-extensions>=4.8.0
requests>=2.31.0
aiohttp>=3.9.1

# File handling
pillow>=10.1.0
python-magic-bin>=0.4.14

# Logging and configuration
loguru>=0.7.2
pyyaml>=6.0.1 